+++ gatsby-node.ts
@@ -1,6 +1,40 @@
+import path from 'path'
+
 import type { GatsbyNode } from 'gatsby'
 
 import { apiSchema } from './src/server'
+
+// Log out information after a build is done
+exports.onPostBuild = ({ reporter }: any) => {
+  reporter.info(`Your Gatsby site has been built!`)
+}
+
+// Create department pages dynamically
+exports.createPages = async ({ graphql, actions }: any) => {
+  const { createPage } = actions
+  const departmentPage = path.resolve(`src/pages/[...slug].tsx`)
+  const result = await graphql(`
+    query {
+      allContentfulPageDepartmentCategory {
+        edges {
+          node {
+            slug
+          }
+        }
+      }
+    }
+  `)
+
+  result.data.allContentfulPageDepartmentCategory.edges.forEach((edge: any) => {
+    createPage({
+      path: `${edge.node.slug}`,
+      component: departmentPage,
+      context: {
+        slug: edge.node.slug,
+      },
+    })
+  })
+}
 
 export const onCreateWebpackConfig: GatsbyNode['onCreateWebpackConfig'] = ({
   actions: { setWebpackConfig },